refactor: 라우터 경로 구조 개선
