+++ apps/client/src/app/App.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { router } from '@router/router';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { RouterProvider } from 'react-router';
@@ -6,7 +7,6 @@ import { RouterProvider } from 'react-router';
 import { ThemeProvider } from '@cds/ui';
 
 import { createQueryClient } from '@shared/libs/query-client';
-import { router } from '@shared/router/router';
 
 function App() {
   const [client] = useState(() => createQueryClient());
