+++ apps/client/src/pages/ai-results/ai-results-page.tsx
@@ -1,11 +1,12 @@
 import { useMemo } from 'react';
 
-import { AiResultsEmptyView } from '@widgets/ai-results-empty-view';
-import { type MockMemo } from '@widgets/memo-list/types/memo';
 import {
   MemoListView,
   type MemoListViewHelpers,
-} from '@widgets/memo-list-view';
+} from '@shared/features/memo-list-view';
+
+import { AiResultsEmptyView } from '@widgets/ai-results-empty-view';
+import { type MockMemo } from '@widgets/memo-list/types/memo';
 
 import { useGetAIMemo } from './apis/queries';
