+++ apps/client/src/pages/all-memo/all-memo-page.tsx
@@ -1,8 +1,9 @@
-import { AllMemoEmptyView } from '@widgets/all-memo-empty-view';
 import {
   MemoListView,
   type MemoListViewHelpers,
-} from '@widgets/memo-list-view';
+} from '@shared/features/memo-list-view';
+
+import { AllMemoEmptyView } from '@widgets/all-memo-empty-view';
 
 import { useGetAllMemo, useGetMemoTotalCount } from './api/queries';
