+++ apps/client/src/router/lazy.ts
