+++ apps/client/src/router/routes/public-route.ts
@@ -1,10 +1,10 @@
+import { PATH } from '@router/path';
 import { RouteObject } from 'react-router';
 
 import { LandingPage } from '@pages/landing';
 import LoginCallbackPage from '@pages/login/login-callback-page';
 
 import { LoginPage } from '../lazy';
-import { PATH } from '../path';
 
 export const publicRoutes: RouteObject[] = [
   {
