+++ apps/client/src/shared/features/ai-prompt/ai-prompt.css.ts
