+++ apps/client/src/shared/features/ai-prompt/components/index.ts
