+++ apps/client/src/shared/features/ai-prompt/components/model/types.ts
