+++ apps/client/src/shared/features/ai-prompt/components/ui/ai-message/ai-message.css.ts
