+++ apps/client/src/shared/features/ai-prompt/components/ui/prompt-header/prompt-header.css.ts
