+++ apps/client/src/shared/features/ai-prompt/components/ui/prompt-input/prompt-input.css.ts
