+++ apps/client/src/shared/features/ai-prompt/components/ui/prompt-option/prompt-option-item.tsx
