+++ apps/client/src/shared/features/ai-prompt/components/ui/prompt-option/prompt-option.css.ts
