+++ apps/client/src/shared/features/ai-prompt/components/ui/selected-memo-section/selected-memo-section.tsx
