+++ apps/client/src/shared/features/ai-prompt/index.ts
@@ -1,0 +1,1 @@
+export { default as AiPrompt } from './ai-prompt';
