+++ apps/client/src/shared/features/memo-list-view/memo-list-view.tsx
@@ -2,7 +2,8 @@ import { AlertModal, FloatingButton } from '@cds/ui';
 
 import { useGetMemoTotalCount } from '@pages/all-memo/api/queries';
 
-import { AiPrompt } from '@widgets/ai-prompt';
+import { AiPrompt } from '@shared/features/ai-prompt';
+
 import { Header } from '@widgets/header';
 import { MemoCardGrid } from '@widgets/memo-list';
 import { type MockMemo } from '@widgets/memo-list/types/memo';
@@ -11,7 +12,7 @@ import { TreeView } from '@widgets/tree-view';
 import {
   type MemoListViewHelpers,
   useMemoListView,
-} from '../hooks/use-memo-list-view';
+} from './hooks/use-memo-list-view';
 
 import * as styles from './memo-list-view.css';
