--- apps/client/src/widgets/ai-prompt/index.ts
@@ -1,1 +1,0 @@
-export { default as AiPrompt } from './ui/ai-prompt';
