+++ apps/client/src/widgets/all-memo-empty-view/ui/all-memo-empty-view.tsx
@@ -1,8 +1,7 @@
+import { PATH } from '@router/path';
 import { useNavigate } from 'react-router';
 
 import { Button } from '@cds/ui';
-
-import { PATH } from '@shared/router/path';
 
 import * as styles from './all-memo-empty-view.css';
