+++ apps/client/vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
     alias: {
       '@app': resolve(__dirname, 'src/app'),
       '@pages': resolve(__dirname, 'src/pages'),
+      '@router': resolve(__dirname, 'src/router'),
       '@widgets': resolve(__dirname, 'src/widgets'),
       '@features': resolve(__dirname, 'src/features'),
       '@entities': resolve(__dirname, 'src/entities'),
